Adding Second Line issue
Hi Professor, I am trying to get the second line to plot on the graph and am having a bit of a problem. I have created a second set of data called "yDes and xDes" and have created a second line (called line 2) but I cannot get it to come up on the plot, I think I am misunderstanding something about the way I am using the class maybe?